# frontend/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "D-WISE Tool Suite Backend API",
    "description": "The REST API for the D-WISE Tool Suite Backend.",
    "version": "beta_mwp_1"
  },
  "paths": {
...
    "/user/{user_id}": {
      "get": {
        "tags": ["user"],
        "summary": "Returns the User with the given ID if it exists",
        "operationId": "get_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "User Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PublicUserRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "patch": {
        "tags": ["user"],
        "summary": "Updates the User with the given ID if it exists",
        "operationId": "update_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "User Id" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "tags": ["user"],
        "summary": "Removes the User with the given ID if it exists",
        "operationId": "delete_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "User Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/project/{proj_id}/metadata": {
      "get": {
        "tags": ["project"],
        "summary": "Returns all ProjectMetadata of the SourceDocument with the given ID if it exists",
        "operationId": "get_all_metadata",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "proj_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Proj Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ProjectMetadataRead" },
                  "title": "Response Project-Get All Metadata"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/sdoc/{sdoc_id}": {
      "get": {
        "tags": ["sourceDocument"],
        "summary": "Returns the SourceDocument with the given ID if it exists",
        "operationId": "get_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "sdoc_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Sdoc Id" } },
          {
            "name": "only_if_finished",
            "in": "query",
            "required": false,
            "schema": { "type": "boolean", "default": true, "title": "Only If Finished" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/SourceDocumentWithDataRead" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "tags": ["sourceDocument"],
        "summary": "Removes the SourceDocument with the given ID if it exists",
        "operationId": "delete_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "sdoc_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Sdoc Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SourceDocumentRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "patch": {
        "tags": ["sourceDocument"],
        "summary": "Updates the SourceDocument with the given ID.",
        "operationId": "update_sdoc",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "sdoc_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Sdoc Id" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SourceDocumentUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SourceDocumentRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/sdoc/{sdoc_id}/metadata": {
      "get": {
        "tags": ["sourceDocument"],
        "summary": "Returns all SourceDocumentMetadata of the SourceDocument with the given ID if it exists",
        "operationId": "get_all_metadata",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "sdoc_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Sdoc Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/SourceDocumentMetadataReadResolved" },
                  "title": "Response Sourcedocument-Get All Metadata"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
    "/sdoc/{sdoc_id}/metadata/{metadata_key}": {
      "get": {
        "tags": ["sourceDocument"],
        "summary": "Returns the SourceDocumentMetadata with the given Key if it exists.",
        "operationId": "read_metadata_by_key",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "sdoc_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Sdoc Id" } },
          {
            "name": "metadata_key",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Metadata Key" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/SourceDocumentMetadataReadResolved" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/doctag/{tag_id}": {
      "get": {
        "tags": ["documentTag"],
        "summary": "Returns the DocumentTag with the given ID.",
        "operationId": "get_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "tag_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Tag Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentTagRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "patch": {
        "tags": ["documentTag"],
        "summary": "Updates the DocumentTag with the given ID.",
        "operationId": "update_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "tag_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Tag Id" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentTagUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentTagRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "tags": ["documentTag"],
        "summary": "Deletes the DocumentTag with the given ID.",
        "operationId": "delete_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "tag_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Tag Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DocumentTagRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/span/{span_id}": {
      "get": {
        "tags": ["spanAnnotation"],
        "summary": "Returns the SpanAnnotation with the given ID.",
        "operationId": "get_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "span_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Span Id" } },
          {
            "name": "resolve",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Resolve Code",
              "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity",
              "default": true
            },
            "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/SpanAnnotationRead" },
                    { "$ref": "#/components/schemas/SpanAnnotationReadResolved" }
                  ],
                  "title": "Response Spanannotation-Get By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "patch": {
        "tags": ["spanAnnotation"],
        "summary": "Updates the SpanAnnotation with the given ID.",
        "operationId": "update_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "span_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Span Id" } },
          {
            "name": "resolve",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Resolve Code",
              "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity",
              "default": true
            },
            "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SpanAnnotationUpdateWithCodeId" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/SpanAnnotationRead" },
                    { "$ref": "#/components/schemas/SpanAnnotationReadResolved" }
                  ],
                  "title": "Response Spanannotation-Update By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "tags": ["spanAnnotation"],
        "summary": "Deletes the SpanAnnotation with the given ID.",
        "operationId": "delete_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "span_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Span Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/SpanAnnotationRead" },
                    { "$ref": "#/components/schemas/SpanAnnotationReadResolved" }
                  ],
                  "title": "Response Spanannotation-Delete By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/spangroup/{span_group_id}": {
      "get": {
        "tags": ["spanGroup"],
        "summary": "Returns the SpanGroup with the given ID.",
        "operationId": "get_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "span_group_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Span Group Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [{ "$ref": "#/components/schemas/SpanGroupRead" }, { "type": "null" }],
                  "title": "Response Spangroup-Get By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "patch": {
        "tags": ["spanGroup"],
        "summary": "Updates the SpanGroup with the given ID.",
        "operationId": "update_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "span_group_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Span Group Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SpanGroupUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [{ "$ref": "#/components/schemas/SpanGroupRead" }, { "type": "null" }],
                  "title": "Response Spangroup-Update By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "tags": ["spanGroup"],
        "summary": "Deletes the SpanGroup with the given ID.",
        "operationId": "delete_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "span_group_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Span Group Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [{ "$ref": "#/components/schemas/SpanGroupRead" }, { "type": "null" }],
                  "title": "Response Spangroup-Delete By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/bbox/{bbox_id}": {
      "get": {
        "tags": ["bboxAnnotation"],
        "summary": "Returns the BBoxAnnotation with the given ID.",
        "operationId": "get_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "bbox_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Bbox Id" } },
          {
            "name": "resolve",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Resolve Code",
              "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity",
              "default": true
            },
            "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/BBoxAnnotationRead" },
                    { "$ref": "#/components/schemas/BBoxAnnotationReadResolvedCode" }
                  ],
                  "title": "Response Bboxannotation-Get By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "patch": {
        "tags": ["bboxAnnotation"],
        "summary": "Updates the BBoxAnnotation with the given ID.",
        "operationId": "update_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "bbox_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Bbox Id" } },
          {
            "name": "resolve",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Resolve Code",
              "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity",
              "default": true
            },
            "description": "If true, the current_code_id of the SpanAnnotation gets resolved and replaced by the respective Code entity"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/BBoxAnnotationUpdateWithCodeId" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/BBoxAnnotationRead" },
                    { "$ref": "#/components/schemas/BBoxAnnotationReadResolvedCode" }
                  ],
                  "title": "Response Bboxannotation-Update By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "tags": ["bboxAnnotation"],
        "summary": "Deletes the BBoxAnnotation with the given ID.",
        "operationId": "delete_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "bbox_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Bbox Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/BBoxAnnotationRead" },
                    { "$ref": "#/components/schemas/BBoxAnnotationReadResolvedCode" }
                  ],
                  "title": "Response Bboxannotation-Delete By Id"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/code/{code_id}": {
      "get": {
        "tags": ["code"],
        "summary": "Returns the Code with the given ID.",
        "operationId": "get_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "code_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Code Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CodeRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "patch": {
        "tags": ["code"],
        "summary": "Updates the Code with the given ID.",
        "operationId": "update_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "code_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Code Id" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CodeUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CodeRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "tags": ["code"],
        "summary": "Deletes the Code with the given ID.",
        "operationId": "delete_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "code_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Code Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CodeRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/memo/{memo_id}": {
      "get": {
        "tags": ["memo"],
        "summary": "Returns the Memo with the given ID if it exists",
        "operationId": "get_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "memo_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Memo Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemoRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "patch": {
        "tags": ["memo"],
        "summary": "Updates the Memo with the given ID if it exists",
        "operationId": "update_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "memo_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Memo Id" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemoUpdate" } } }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemoRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      },
      "delete": {
        "tags": ["memo"],
        "summary": "Removes the Memo with the given ID if it exists",
        "operationId": "delete_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          { "name": "memo_id", "in": "path", "required": true, "schema": { "type": "integer", "title": "Memo Id" } }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MemoRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
    "/sdocmeta": {
      "put": {
        "tags": ["sdocMetadata"],
        "summary": "Creates a new Metadata and returns it with the generated ID.",
        "operationId": "create_new_metadata",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SourceDocumentMetadataCreate" } }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/SourceDocumentMetadataRead" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
...
    "/feedback/{feedback_id}": {
      "get": {
        "tags": ["feedback"],
        "summary": "Returns the Feedback with the given ID.",
        "operationId": "get_by_id",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "feedback_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Feedback Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FeedbackRead" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": { "items": { "$ref": "#/components/schemas/ValidationError" }, "type": "array", "title": "Detail" }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicUserRead": {
        "properties": {
          "id": { "type": "integer", "title": "Id", "description": "ID of the User" },
          "first_name": { "type": "string", "title": "First Name", "description": "First name of the User" },
          "last_name": { "type": "string", "title": "Last Name", "description": "Last name of the User" }
        },
        "type": "object",
        "required": ["id", "first_name", "last_name"],
        "title": "PublicUserRead",
        "description": "A user object with information that everybody may see."
      },
...
      "UserUpdate": {
        "properties": {
          "email": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Email",
            "description": "E-Mail of the User"
          },
          "first_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "First Name",
            "description": "First name of the User"
          },
          "last_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Last Name",
            "description": "Last name of the User"
          },
          "password": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Password",
            "description": "Hashed password of the User"
          }
        },
        "type": "object",
        "title": "UserUpdate"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "/authentication/login" } }
      }
    }
  }
}